fix potential invalid buffer access with negative encoding values
git-svn-id: svn://scm.orgis.org/mpg123/trunk@1910 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/id3.c
@@ -186,14 +186,14 @@ void id3_link(mpg123_handle *fr)
 */
 void store_id3_text(mpg123_string *sb, char *source, size_t source_size, const int noquiet)
 {
-	int encoding;
+	unsigned int encoding;
 	unsigned int bwidth;
 	if(!source_size)
 	{
 		debug("Empty id3 data!");
 		return;
 	}
-	encoding = source[0];
+	encoding = (unsigned int) source[0];
 	++source;
 	--source_size;
 	debug1("encoding: %i", encoding);
